backend/requirements.txt
# Core Django & API
Django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
djangorestframework-simplejwt==5.2.2
django-environ==0.10.0

# Database
psycopg2-binary==2.9.7

# Task Queue & Messaging
celery==5.3.1
redis==4.6.0
kombu==5.3.1

# Web Scraping
Scrapy==2.10.1
scrapy-user-agents==0.1.1
scrapy-rotating-proxies==0.6.2

# Data Processing & ML
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6

# gRPC
grpcio==1.56.2
grpcio-tools==1.56.2
protobuf==4.23.4

# Real-time WebSockets
channels==4.0.0
channels-redis==4.1.0

# Monitoring & Metrics
prometheus-client==0.17.1
django-prometheus==2.3.1

# Workflow Orchestration
apache-airflow==2.6.3
apache-airflow-providers-postgres==5.5.1

# Development & Production
gunicorn==21.2.0
uvicorn==0.23.2
whitenoise==6.5.0
python-decouple==3.8